feat: created django app for flask app deployment
