+++ BackEndApp/BackEndApp/urls.py
@@ -20,4 +20,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("v1/", include("v1.urls")),
     path("auth/", include("authv1.urls")),
+    path("deployments/", include("deployments.urls")),
 ]
