+++ BackEndApp/v1/views.py
@@ -1,16 +1,12 @@
 from genericpath import isdir
 from django.shortcuts import render
-from django.http import JsonResponse, HttpResponse, FileResponse
+from django.http import JsonResponse, HttpResponse
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 import sys
 import os
 import json
 import importlib
-import stat
-import git
-import zipfile
-import shutil
 
 from authv1.store import Store
 from authv1.models import User
@@ -707,396 +703,4 @@ def share_project(request):
             )
     except:
         status, success, message = 500, False, "Failed"
-    return JsonResponse({"success": success, "message": message}, status=status)
-
-
-@api_view(["POST"])
-@is_authenticated
-def local_deploy(request):
-    """
-    Endpoint for locally deploying user model.
-    Inputs:
-    -------
-    request:
-        Requires: username, project_id, model_categories & model_download_type in request data
-                  token in request header
-    Returns:
-    --------
-    response: HttpResponse (Required Download)
-              (Yes, not JsonResponse like other views)
-    """
-
-    def retrieve_file_paths(dirName):
-        filePaths = []
-
-        # Read all directory, subdirectories and file lists
-        for root, directories, files in os.walk(dirName):
-            for filename in files:
-                # Create the full filepath by using os module.
-                filePath = os.path.join(root, filename)
-                filePaths.append(filePath)
-
-        # return all paths
-        return filePaths
-
-    try:
-        username = request.data.get("username")
-        user = User(username=username, password=None)
-        user = user.find()
-
-        store_obj = Store(user)
-        project_id = request.data.get("project_id")
-
-        if not store_obj.exist(project_id):
-            raise Exception("No such project exists")
-
-        project_dir = store_obj.path + os.sep + project_id
-
-        model_categories = request.data.get("model_categories")
-        model_download_type = request.data.get("model_download_type")
-
-        deployment_dir = os.path.join(project_dir, "deployment")
-        flask_app_url = os.getenv("FLASK_APP_HTTPS_URL")
-
-        if os.path.exists(deployment_dir) and os.path.isdir(deployment_dir):
-            for root, dirs, files in os.walk(deployment_dir, topdown=False):
-                for name in files:
-                    filename = os.path.join(root, name)
-                    os.chmod(filename, stat.S_IWUSR)
-                    os.remove(filename)
-                for name in dirs:
-                    os.rmdir(os.path.join(root, name))
-            os.rmdir(deployment_dir)
-
-        git.Repo.clone_from(f"{flask_app_url}", f"{deployment_dir}", branch="main")
-
-        program_path = os.path.join(deployment_dir, "app.py")
-        modified_flask_lines = []
-
-        with open(program_path, "r") as flask_fh:
-            flask_lines = flask_fh.readlines()
-
-            categories = "CATEGORIES = []\n"
-            modified_categories = f"CATEGORIES = {model_categories}\n"
-
-            for line in flask_lines:
-                if line == categories:
-                    modified_flask_lines.append(modified_categories)
-                else:
-                    modified_flask_lines.append(line)
-
-        with open(program_path, "w") as flask_fh:
-            flask_fh.writelines(modified_flask_lines)
-
-        if model_download_type == "zip":
-            current_dir = os.getcwd()
-            os.chdir(deployment_dir)
-
-            filePaths = retrieve_file_paths(".")
-            zip_dir = os.path.join(project_dir, "deployment.zip")
-            zip_file = zipfile.ZipFile(zip_dir, "w")
-            with zip_file:
-                for file in filePaths:
-                    zip_file.write(file)
-
-            os.chdir(current_dir)
-
-            response = HttpResponse(
-                open(zip_dir, "rb"),
-                headers={
-                    "Content-Type": "application/zip",
-                    "Content-Disposition": "attachment; filename=deployment.zip",
-                },
-            )
-
-            if os.path.exists(deployment_dir) and os.path.isdir(deployment_dir):
-                for root, dirs, files in os.walk(deployment_dir, topdown=False):
-                    for name in files:
-                        filename = os.path.join(root, name)
-                        os.chmod(filename, stat.S_IWUSR)
-                        os.remove(filename)
-                    for name in dirs:
-                        os.rmdir(os.path.join(root, name))
-                os.rmdir(deployment_dir)
-
-            os.remove(zip_dir)
-
-            return response
-
-        else:
-            status, success, message = 200, True, "Executable Downloaded Successfully"
-            return JsonResponse({"success": success, "message": message}, status=status)
-
-    except Exception as e:
-        status, success, message = 500, False, "Deployment Attempt Failed"
-        return JsonResponse({"success": success, "message": message}, status=status)
-
-
-@api_view(["POST"])
-@is_authenticated
-def cloud_deploy(request):
-    """
-    Endpoint for deploying user model to user-preferred cloud environment (AWS / GCP).
-    Inputs:
-    -------
-    request:
-        Requires: username, project_id, model_categories & pickle file details in request data
-                  token in request header
-    Returns:
-    --------
-    response: JsonResponse for Cloud Deployment describing Success or Failure
-    """
-
-    try:
-        username = request.data.get("username")
-        user = User(username=username, password=None)
-        user = user.find()
-
-        store_obj = Store(user)
-        project_id = request.data.get("project_id")
-
-        if not store_obj.exist(project_id):
-            raise Exception("No such project exists")
-
-        project_dir = store_obj.path + os.sep + project_id
-
-        model_categories = request.data.get("model_categories")
-
-        # Append pickle contents in chunks
-        current_chunk = request.data.get("current_chunk")
-        total_chunks = request.data.get("total_chunks")
-        pkl_file_bytes = request.data.get("pkl_file_bytes")
-        pkl_file_content = bytearray(pkl_file_bytes.values())
-
-        pkl_dir = os.path.join(project_dir, "pickle")
-        pkl_path = os.path.join(pkl_dir, f"model.pkl")
-
-        if os.path.exists(pkl_dir) and os.path.isdir(pkl_dir):
-            with open(pkl_path, "ab") as pkl_fh:
-                pkl_fh.write(pkl_file_content)
-        else:
-            os.mkdir(pkl_dir)
-            with open(pkl_path, "wb") as pkl_fh:
-                pkl_fh.write(pkl_file_content)
-
-        if current_chunk == total_chunks:
-            # Clone Flask App once upload complete
-            deployment_dir = os.path.join(project_dir, "deployment")
-            flask_app_url = os.getenv("FLASK_APP_HTTPS_URL")
-
-            if os.path.exists(deployment_dir) and os.path.isdir(deployment_dir):
-                for root, dirs, files in os.walk(deployment_dir, topdown=False):
-                    for name in files:
-                        filename = os.path.join(root, name)
-                        os.chmod(filename, stat.S_IWUSR)
-                        os.remove(filename)
-                    for name in dirs:
-                        os.rmdir(os.path.join(root, name))
-                os.rmdir(deployment_dir)
-
-            git.Repo.clone_from(f"{flask_app_url}", f"{deployment_dir}", branch="main")
-
-            new_pkl_path = os.path.join(deployment_dir, "model.pkl")
-
-            shutil.copy(pkl_path, new_pkl_path)
-
-            if os.path.exists(pkl_dir) and os.path.isdir(pkl_dir):
-                for root, dirs, files in os.walk(pkl_dir, topdown=False):
-                    for name in files:
-                        filename = os.path.join(root, name)
-                        os.chmod(filename, stat.S_IWUSR)
-                        os.remove(filename)
-                    for name in dirs:
-                        os.rmdir(os.path.join(root, name))
-                os.rmdir(pkl_dir)
-
-            program_path = os.path.join(deployment_dir, "app.py")
-            modified_flask_lines = []
-
-            with open(program_path, "r") as flask_fh:
-                flask_lines = flask_fh.readlines()
-
-                model_path = 'MODEL_PATH = ""\n'
-                categories = "CATEGORIES = []\n"
-
-                modified_model_path = 'MODEL_PATH = "./model.pkl"\n'
-                modified_categories = f"CATEGORIES = {model_categories}\n"
-
-                for line in flask_lines:
-                    if line == model_path:
-                        modified_flask_lines.append(modified_model_path)
-                    elif line == categories:
-                        modified_flask_lines.append(modified_categories)
-                    else:
-                        modified_flask_lines.append(line)
-
-            with open(program_path, "w") as flask_fh:
-                flask_fh.writelines(modified_flask_lines)
-
-            status, success, message = 200, True, "Cloud Deployment Successful"
-        else:
-            status, success, message = 204, True, "Cloud Deployment Underway"
-
-    except Exception as e:
-        status, success, message = 500, False, "Deployment Attempt Failed"
-
-    return JsonResponse({"success": success, "message": message}, status=status)
-
-
-@api_view(["POST"])
-@is_authenticated
-def hybrid_deploy(request):
-    """
-    Endpoint for deploying user model in both local and cloud environments.
-    Inputs:
-    -------
-    request:
-        Requires: username, project_id, model_categories, download_type & pickle file details in request data
-                  token in request header
-    Returns:
-    --------
-    response: HttpResponse (Required Download) for Local Deployment
-              JsonResponse for Cloud Deployment describing Success or Failure
-    """
-
-    def retrieve_file_paths(dirName):
-        filePaths = []
-
-        # Read all directory, subdirectories and file lists
-        for root, directories, files in os.walk(dirName):
-            for filename in files:
-                # Create the full filepath by using os module.
-                filePath = os.path.join(root, filename)
-                filePaths.append(filePath)
-
-        # return all paths
-        return filePaths
-
-    try:
-        username = request.data.get("username")
-        user = User(username=username, password=None)
-        user = user.find()
-
-        store_obj = Store(user)
-        project_id = request.data.get("project_id")
-
-        if not store_obj.exist(project_id):
-            raise Exception("No such project exists")
-
-        project_dir = store_obj.path + os.sep + project_id
-
-        # Append pickle contents in chunks
-        current_chunk = request.data.get("current_chunk")
-        total_chunks = request.data.get("total_chunks")
-        pkl_file_bytes = request.data.get("pkl_file_bytes")
-        pkl_file_content = bytearray(pkl_file_bytes.values())
-
-        pkl_dir = os.path.join(project_dir, "pickle")
-        pkl_path = os.path.join(pkl_dir, f"model.pkl")
-
-        if os.path.exists(pkl_dir) and os.path.isdir(pkl_dir):
-            with open(pkl_path, "ab") as pkl_fh:
-                pkl_fh.write(pkl_file_content)
-        else:
-            os.mkdir(pkl_dir)
-            with open(pkl_path, "wb") as pkl_fh:
-                pkl_fh.write(pkl_file_content)
-
-        if current_chunk == total_chunks:
-            # Clone Flask App once upload complete
-            deployment_dir = os.path.join(project_dir, "deployment")
-            flask_app_url = os.getenv("FLASK_APP_HTTPS_URL")
-
-            if os.path.exists(deployment_dir) and os.path.isdir(deployment_dir):
-                for root, dirs, files in os.walk(deployment_dir, topdown=False):
-                    for name in files:
-                        filename = os.path.join(root, name)
-                        os.chmod(filename, stat.S_IWUSR)
-                        os.remove(filename)
-                    for name in dirs:
-                        os.rmdir(os.path.join(root, name))
-                os.rmdir(deployment_dir)
-
-            git.Repo.clone_from(f"{flask_app_url}", f"{deployment_dir}", branch="main")
-
-            new_pkl_path = os.path.join(deployment_dir, "model.pkl")
-            shutil.copy(pkl_path, new_pkl_path)
-
-            if os.path.exists(pkl_dir) and os.path.isdir(pkl_dir):
-                for root, dirs, files in os.walk(pkl_dir, topdown=False):
-                    for name in files:
-                        filename = os.path.join(root, name)
-                        os.chmod(filename, stat.S_IWUSR)
-                        os.remove(filename)
-                    for name in dirs:
-                        os.rmdir(os.path.join(root, name))
-                os.rmdir(pkl_dir)
-
-            model_categories = request.data.get("model_categories")
-            model_download_type = request.data.get("model_download_type")
-
-            program_path = os.path.join(deployment_dir, "app.py")
-            modified_flask_lines = []
-
-            with open(program_path, "r") as flask_fh:
-                flask_lines = flask_fh.readlines()
-
-                model_path = 'MODEL_PATH = ""\n'
-                categories = "CATEGORIES = []\n"
-
-                modified_model_path = 'MODEL_PATH = "./model.pkl"\n'
-                modified_categories = f"CATEGORIES = {model_categories}\n"
-
-                for line in flask_lines:
-                    if line == model_path:
-                        modified_flask_lines.append(modified_model_path)
-                    elif line == categories:
-                        modified_flask_lines.append(modified_categories)
-                    else:
-                        modified_flask_lines.append(line)
-
-            with open(program_path, "w") as flask_fh:
-                flask_fh.writelines(modified_flask_lines)
-
-            if model_download_type == "zip":
-                current_dir = os.getcwd()
-                os.chdir(deployment_dir)
-
-                filePaths = retrieve_file_paths(".")
-                zip_dir = os.path.join(project_dir, "deployment.zip")
-                zip_file = zipfile.ZipFile(zip_dir, "w")
-                with zip_file:
-                    for file in filePaths:
-                        zip_file.write(file)
-
-                os.chdir(current_dir)
-
-                response = HttpResponse(
-                    open(zip_dir, "rb"),
-                    headers={
-                        "Content-Type": "application/zip",
-                        "Content-Disposition": "attachment; filename=deployment.zip",
-                    },
-                )
-
-                os.remove(zip_dir)
-
-                return response
-
-            else:
-                status, success, message = (
-                    200,
-                    True,
-                    "Executable Downloaded Successfully",
-                )
-                return JsonResponse(
-                    {"success": success, "message": message}, status=status
-                )
-
-        else:
-            status, success, message = 204, True, "Hybrid Deployment Underway"
-
-    except Exception as e:
-        status, success, message = 500, False, "Deployment Attempt Failed"
-
     return JsonResponse({"success": success, "message": message}, status=status)
